move MapperFactory.ts to the impl folder in the visual style model
